pyproject.toml
[project]
name = "maestro"
version = "0.1.0"
description = "Intelligent LLM orchestration platform"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.115",
    "python-multipart>=0.0.7",
    "uvicorn[standard]>=0.34",
    "pyyaml>=6.0",
]
openai = ["openai>=1.82"]
anthropic = ["anthropic>=0.52"]
google = ["google-genai>=1.0"]
ollama = ["ollama>=0.4"]
chroma = ["chromadb>=0.6"]
grpc = ["grpcio>=1.60"]
qdrant = ["qdrant-client>=1.12"]
all = ["maestro[server,openai,anthropic,google,ollama,chroma,grpc]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.11",
]

[project.scripts]
maestro = "maestro.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/maestro"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
